Monero confirmations are a u64
Trying to deserialize the number as a u32 caused deserialization
errors.

+++ monero-rpc/src/rpc/wallet.rs
@@ -470,7 +470,7 @@ struct CheckTxKeyParams {
 
 #[derive(Clone, Copy, Debug, Deserialize)]
 pub struct CheckTxKey {
-    pub confirmations: u32,
+    pub confirmations: u64,
     pub received: u64,
 }
